# src/main/resources/static/api.yaml
openapi: "3.0.0"
info:
  version: "1"
  title: darfichraus.de API
servers:
  - url: https://api.darfichraus.de
    description: production instance
  - url: https://api.dev.crimsy.tech
    description: development instance
security:
  - ApiKeyAuth: []
...
paths:
...
  /admin/restrictions:
    get:
      tags:
        - admin
      security:
        - BearerAuth: []
      operationId: getAllRestrictions
      summary: Get all published restrictions
      responses:
        '200':
          description: List of all restrictions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restrictions'
    put:
      tags:
        - admin
      security:
        - BearerAuth: []
      operationId: updateRestriction
      summary: updates a restriction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Restriction'
      responses:
        '204':
          description: successfully updated Restriction
  /admin/restrictions/{id}:
    delete:
      tags:
        - admin
      security:
        - BearerAuth: []
      operationId: deleteRestriction
      summary: deletes a restriction
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '204':
          description: successfully deleted Restriction
  /admin/additionalInformation:
    post:
      tags:
        - admin
      security:
        - BearerAuth: []
      operationId: addAdditionalInformationToCategory
      parameters:
        - in: path
          name: category
          schema:
            type: string
          required: true
          example: 'Bayern'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdditionalInformation'
      responses:
        '200':
          description: successfully added additional information
  /admin/subscriptions/{areal}/{arealIdentifier}:
    get:
      tags:
        - admin
      operationId: getSubscriptionsByArealAndArealIdentifier
      summary: Get subscriptions for specific areal and arealIdentifier
      parameters:
        - in: path
          name: areal
          schema:
            $ref: '#/components/schemas/Areal'
          required: true
          example: 'STATE'
        - in: path
          name: arealIdentifier
          schema:
            type: string
          required: true
          example: 'Bayern'
      responses:
        '200':
          description: Subscriptions for given areal and arealIdentifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriptions'
...
  /admin/users:
    get:
      tags:
        - userManagement
      security:
        - BearerAuth: []
      operationId: getProfiles
      summary: gets all users
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /admin/users/{email}:
    delete:
      tags:
        - userManagement
      security:
        - BearerAuth: []
      operationId: deleteUser
      parameters:
        - in: path
          name: email
          schema:
            type: string
            format: email
          required: true
          example: 'te@st.de'
      responses:
        '204':
          description: successfully deleted user
  /admin/users/passwordReset:
    patch:
      tags:
        - userManagement
      security:
        - BearerAuth: []
      operationId: changePassword
      summary: change User password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '204':
          description: successfully deleted user
  /admin/users/register:
    post:
      tags:
        - userManagement
      security:
        - BearerAuth: []
      operationId: registerUser
      summary: adds a user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsWithRoles'
      responses:
        '204':
          description: successfully added user
  /admin/users/editUser:
    post:
      tags:
        - userManagement
      security:
        - BearerAuth: []
      operationId: editUser
      summary: edits a user profile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '204':
          description: successfully added user
...
  /admin/situation-advisor/situation-types:
    post:
      tags:
        - situationTypes_Admin
      security:
        - BearerAuth: []
      operationId: addSituationType
      summary: adds a new Situation Type
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SituationType'
      responses:
        '204':
          description: Situation Type successfully added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SituationType'
    put:
      tags:
        - situationTypes_Admin
      security:
        - BearerAuth: []
      operationId: updateSituationType
      summary: Updates a Situation Type
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SituationType'
      responses:
        '204':
          description: Situation Type successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SituationType'
  /admin/situation-advisor/situation-types/{id}:
    delete:
      tags:
        - situationTypes_Admin
      security:
        - BearerAuth: []
      operationId: deleteSituationType
      summary: Deletes a Situation Type
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Situation Type successfully deleted
...
  /admin/situation-advisor/situations:
    post:
      tags:
        - situations_Admin
      security:
        - BearerAuth: []
      operationId: addSituation
      summary: Creates a new situation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Situation'
      responses:
        '204':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Situation'

    put:
      tags:
        - situations_Admin
      security:
        - BearerAuth: []
      operationId: updateSituation
      summary: Updates a situation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Situation'
      responses:
        '204':
          description: Update of Situation successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Situation'
  /admin/situation-advisor/situations/{id}:
    delete:
      tags:
        - situations_Admin
      security:
        - BearerAuth: []
      operationId: deleteSituation
      summary: Deletes a Situation
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Situation successfully deleted
...
  /admin/situation-advisor/situations-message-types:
    post:
      tags:
        - situationMessageTypes_Admin
      security:
        - BearerAuth: []
      operationId: addSituationMessageType
      summary: Creates a new Situation Message Type
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SituationMessageType'
      responses:
        '204':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SituationMessageType'
    put:
      tags:
        - situationMessageTypes_Admin
      security:
        - BearerAuth: []
      operationId: updateSituationMessageType
      summary: Updates a Situation Message Type
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SituationMessageType'
      responses:
        '204':
          description: Update of Situation Message Type successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SituationMessageType'
  /admin/situation-advisor/situation-message-types/{id}:
    delete:
      tags:
        - situationMessageTypes_Admin
      security:
        - BearerAuth: []
      operationId: deleteSituationMessageType
      summary: Deletes a Situation Message Type
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Situation Message Type successfully
...
  /admin/situation-advisor/situations-messages:
    post:
      tags:
        - situationMessages_Admin
      security:
        - BearerAuth: []
      operationId: addSituationMessage
      summary: Creates a new Situation Message
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SituationMessage'
      responses:
        '204':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SituationMessage'

    put:
      tags:
        - situationMessages_Admin
      security:
        - BearerAuth: []
      operationId: updateSituationMessage
      summary: Updates a Situation Message
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SituationMessage'
      responses:
        '204':
          description: Update of Situation Message successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SituationMessage'
  /admin/situation-advisor/situation-messages/{id}:
    delete:
      tags:
        - situationMessages_Admin
      security:
        - BearerAuth: []
      operationId: deleteSituationMessage
      summary: Deletes a Situation Message
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Situation Message successfully
...
  /admin/situation-advisor/situations-references:
    post:
      tags:
        - situationReferences_Admin
      security:
        - BearerAuth: []
      operationId: addSituationReference
      summary: Creates a new Situation Reference
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SituationReference'
      responses:
        '204':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SituationReference'

    put:
      tags:
        - situationReferences_Admin
      security:
        - BearerAuth: []
      operationId: updateSituationReference
      summary: Updates a Situation Reference
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SituationReference'
      responses:
        '204':
          description: Update of Situation Message successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SituationReference'
  /admin/situation-advisor/situation-references/{id}:
    delete:
      tags:
        - situationReferences_Admin
      security:
        - BearerAuth: []
      operationId: deleteSituationReference
      summary: Deletes a Situation Reference
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Situation Reference successfully
components:
  schemas:
    Restriction:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          required:
            - areal
            - arealIdentifier
            - restrictionState
            - restrictionType
            - restrictionStart
            - restrictionEnd
            - recipient
            - publisher
            - restrictionDescription
            - shortDescription
          properties:
            areal:
              $ref: '#/components/schemas/Areal'
            arealIdentifier:
              type: string
            restrictionType:
              $ref: '#/components/schemas/RestrictionType'
            restrictionState:
              $ref: '#/components/schemas/RestrictionState'
            shortDescription:
              type: string
              description: 'Give a short excerpt, why the restriction is active'
              example: 'There are zombies everywhere'
            restrictionDescription:
              type: string
              description: 'Full styled text to give a closer look to the restriction.'
            restrictionStart:
              type: string
              format: date
              description: 'Start of restriction'
              example: '2020-03-21'
            restrictionEnd:
              type: string
              format: date
              description: 'End of restriction'
              example: '2020-03-21'
            recipient:
              type: string
              description: 'affected group of citizens'
              example: 'Schüler'
            publisher:
              type: string
              description: 'source of information'
              example: 'Regierung'
            furtherInformation:
              type: string
              description: 'Further information for snoopy people'
              example: 'http://link/to/more/information'
            verified:
              type: boolean
              description: 'whether this restriction has been verified yet'
              default: false
    Restrictions:
      type: array
      items:
        $ref: '#/components/schemas/Restriction'
    Areal:
      description: 'A restriction can be city-based (zip), county based (Landkreis), for the whole state (Bayern) or countrywide (country)'
      type: string
      enum:
        - ZIP
        - COUNTY
        - STATE
        - COUNTRY
    Role:
      description: 'Roles used in the app'
      type: string
      enum:
        - ROLE_ADMIN
        - ROLE_COLLABORATOR
        - ROLE_USER
    CredentialsWithRoles:
      allOf:
        - $ref: '#/components/schemas/Credentials'
        - type: object
          required:
            - roles
          properties:
            roles:
              type: array
              items:
                $ref: '#/components/schemas/Role'
    Credentials:
      type: object
      description: 'user credentials for login'
      required:
        - username
        - password
      properties:
        username:
          type: string
          format: email
        password:
          type: string
          format: password
    User:
      type: object
      description: 'user info object'
      required:
        - username
        - roles
      properties:
        username:
          type: string
          format: email
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
    RestrictionType:
      description: 'A restriction can affect one of these  areas of public life'
      type: string
      enum:
        - GENERAL_INFORMATION
        - PUBLIC_TRANSPORTATION
        - EVENTS_AND_ASSEMBLIES
        - GASTRONOMY
        - PUBLIC_PLACES
        - RETAIL
        - CURFEW
    RestrictionState:
      description: 'Describes severeness of restriction'
      type: string
      enum:
        - BAN
        - RESTRICTION
...
    AdditionalInformation:
      type: object
      additionalProperties:
        type: string
    Subscriptions:
      type: array
      items:
        $ref: '#/components/schemas/Subscription'
    Subscription:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          required:
            - email
            - areal
            - arealIdentifier
            - contactAllowed
          properties:
            email:
              type: string
              example: 'a@b.com'
            areal:
              $ref: '#/components/schemas/Areal'
            arealIdentifier:
              type: string
            types:
              type: array
              items:
                $ref: '#/components/schemas/RestrictionType'
    SituationType:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - required:
            - name
            - description
            - severity
        - type: object
          description: "This object allows to categorize Situations into high level crysis."
          properties:
            name:
              type: string
              description: "Classifier."
              example: "Pandemie"
            description:
              type: string
              example: "Ausbruch einer Krankheit. Betrifft den gesamten Globus"
            severity:
              $ref: '#/components/schemas/SituationSeverity'
            media:
              type: string
              example: "virus.png"
    Situation:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          description: "A Situation is a current crysis."
          required:
            - name
            - startDate
            - severity
          properties:
            name:
              type: string
              description: "Describes the situation; like a title"
              example: "Corona/Covid-19"
            startDate:
              type: string
              format: date
              description: "Start of the situation/crysis"
              example: "2020-01-01"
            endDate:
              type: string
              format: date
              nullable: true
              description: "End of situation/crysis"
              example: "2020-12-31"
            severity:
              $ref: '#/components/schemas/SituationSeverity'
            active:
              type: boolean
              default: true
              description: "Toggles the afffectness of a situation/crysis. False will not trigger any action to a system/user (e. g. push notification)"
    SituationMessageType:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          description: "Classifies a SituationMessage into a type. Has impact to icon, coloring, severity of messages their handling"
          required:
            - name
            - severity
            - color
          properties:
            name:
              type: string
              description: "Speakable type of Type"
              example: "Restrictive Instruktion"
            media:
              type: string
              description: "Icon"
              example: "exclamation-mark.svg"
            severity:
              $ref: '#/components/schemas/SituationSeverity'
            color:
              type: string
              description: "Provides the possibility to annotate messages/headers with MessageType-based color"
              example: "orange"
    SituationMessage:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          description: "Represents an instruction or an information to recipients or systems."
          required:
            - situationId
            - messageTypeId
            - title
            - message
            - abstract
            - severity
          properties:
            situationId:
              type: string
              description: "Sorts the message to a situation/crysis"
              example: "5eaf148a0d077d75108087d3"
            messageTypeId:
              type: string
              description: "Classifies the Message as type of e.g. Restrictive-Instruction; inherits color schema, severity, etc."
              example: "5eaf17692c1f4e9671ec7111"
            title:
              type: string
              description: "Quick information/abstract about the message"
              example: "Wearing a face mask is mandatory!"
            message:
              type: string
              description: "Whole information. Fully formatable."
              example: "By start of April, 20th 2020 you have to wear a face mask in supermarkets and all shop-alike stores."
            abstract:
              type: string
              description: "Short text (excerpt) from the whole message"
              example: "Please wear a face mask in shopping stores"
            icon:
              type: string
              description: "Icon"
              example: "mask.svg"
            severity:
              $ref: '#/components/schemas/SituationSeverity'
            documents:
              type: array
              items:
                $ref: '#/components/schemas/SituationReference'
            areaOfEffect:
              type: array
              description: "A list of all GeoJson-Ids, where the Situation is taking place/is affected"
              items:
                type: string
            status:
              $ref: '#/components/schemas/SituationMessageStatus'
            version:
              type: number
              description: "Every change increments the version. Full history possible"
    SituationReference:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          description: "References to source of information"
          properties:
            title:
              type: string
              description: "Allows to describe the source/reference"
              example: "How to wear a face mask?"
            type:
              $ref: '#/components/schemas/SituationReferenceType'
            classifier:
              $ref: '#/components/schemas/SituationReferenceClassifier'
            source:
              type: string
              description: "points to the source"
              example: "http://darfichraus.de/mask-mandatory.pdf"
    SituationReferenceType:
      description: "Type of reference to build type-based-action"
      type: string
      enum:
        - URL
        - FILE
        - PICTURE
        - VIDEO
        - VOICE
        - OTHER
    SituationReferenceClassifier:
      description: "Classifies the reference as type of"
      type: string
      enum:
        - ATTACHMENT
        - SOURCE
        - OTHER
    SituationMessageStatus:
      description: "Get information about the status"
      type: string
      enum:
        - DRAFT
        - APPROVED
        - REVOKED
        - OTHER
    SituationSeverity:
      description: "Describes the severity of a subject"
      type: number
      enum:
        - 1
        - 2
        - 3
        - 4
        - 5
    BaseEntity:
      type: object
      discriminator:
        propertyName: Entity
      properties:
        id:
          type: string
        modified:
          type: string
          format: date-time
          description: 'when this entity has been last modified'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: 'API-KEY'
      description: 'Api-Key needed to authenticate requests'
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT